To create an auto color change effect from right to left using CSS, you can use CSS animations and keyframes. Here's an example code that demonstrates this effect:
In this example, we have a div with the class "color-changing-box." We use CSS animations to create a color change effect from red to blue, starting from the right side of the box to the left. The animation duration is set to 5 seconds, and it loops infinitely using the "infinite" keyword.

You can adjust the width, height, colors, animation duration, and other styles to customize the effect to your liking. Save the HTML and CSS code in separate files with the specified names ("index.html" and "styles.css") and place them in the same directory. When you open the "index.html" file in a web browser, you should see the color-changing box with the auto color change effect.
https://kolkataffghoshbabu.net/